Worker: Add tests for worker creation in data URL iframe and dedicated worker
This CL adds tests for dedicated/shared worker creation in data URL
contexts (iframe and dedicated worker) whose origins are opaque.

For dedicated worker creation:
  1. Creating a dedicated worker in a data URL iframe
  2. Creating a dedicated worker in a data URL dedicated worker
  3. Creating a data URL dedicated worker in a data URL iframe
  4. Creating a data URL dedicated worker in a data URL dedicated worker

1 and 3 are expected to asynchronously fail worker construction because
the same-origin policy is supposed to be checked during the fetch
algorithm based on request's "same-origin" mode. However, Chrome
synchronously fails the cases because the same-origin policy is checked
in the Worker constructor (see SecurityOrigin::CanReadContent() call in
AbstractWorker::ResolveURL()). See issue 1055697.

For shared workers:
  5. Creating a shared worker in a data URL iframe
  6. Creating a data URL shared worker in a data URL iframe

5 fails because of the same reason as 1 and 3. 6 should pass like 2 and
4, but fails because of issue 1055693. Dedicated workers cannot create
shared workers unlike 2 and 4, so there are no tests for the case.

Bug: 1053382, 1055693, 1055697
Change-Id: I8480b6079d4fb271855656cedcffc86e6e6c0fc8